rename: socketEvents -> sockets
